Move HiddenView view asserion to espresso package
