+++ app/src/androidTest/java/com/twilio/video/app/espresso/HiddenView.kt
@@ -1,4 +1,4 @@
-package com.twilio.video.app
+package com.twilio.video.app.espresso
 
 import android.view.View
 import androidx.test.espresso.NoMatchingViewException
